Move file helper methods to own submodule
These are all methods which have something to do with manipulating files or extracting file contents, not the actual process of ingestion. Move them out to their own submodule to help make boundaries clearer.
